build(composer): increase composer's base PHP requirement for Nullable types
Type declarations for parameters and return values can now be marked as nullable
by prefixing the type name with a question mark. This signifies that as well as
the specified type, null can be passed as an argument, or returned as a value,
respectively. Now used for Data Structures.

The CLI command:
- `$ composer require php:^7.1`

More info:
[PHP.net - PHP7.1 New features](https://www.php.net/manual/en/migration71.new-features.php#migration71.new-features.nullable-types).